Move .NET resources read/writer code to dnlib
