+++ de4dot.code/de4dot.code.csproj
@@ -338,16 +338,6 @@
     <Compile Include="renamer\TypeNames.cs" />
     <Compile Include="renamer\TypeRenamerState.cs" />
     <Compile Include="renamer\VariableNameState.cs" />
-    <Compile Include="resources\BuiltInResourceData.cs" />
-    <Compile Include="resources\IResourceData.cs" />
-    <Compile Include="resources\ResourceDataCreator.cs" />
-    <Compile Include="resources\ResourceElement.cs" />
-    <Compile Include="resources\ResourceElementSet.cs" />
-    <Compile Include="resources\ResourceReader.cs" />
-    <Compile Include="resources\ResourceTypeCode.cs" />
-    <Compile Include="resources\ResourceWriter.cs" />
-    <Compile Include="resources\UserResourceData.cs" />
-    <Compile Include="resources\UserResourceType.cs" />
     <Compile Include="SharpZipLib\Checksums\Adler32.cs" />
     <Compile Include="SharpZipLib\Checksums\IChecksum.cs" />
     <Compile Include="SharpZipLib\Main.cs" />
