+++ de4dot.code/deobfuscators/CodeVeil/ResourceConverter.cs
@@ -18,20 +18,33 @@
 */
 
 using System;
+using System.Drawing;
 using System.IO;
+using System.Runtime.Serialization;
 using dnlib.IO;
 using dnlib.DotNet;
-using de4dot.code.resources;
+using dnlib.DotNet.Resources;
 
 namespace de4dot.code.deobfuscators.CodeVeil {
 	class ResourceConverter {
 		ModuleDefMD module;
 		ResourceInfo[] infos;
-		ResourceDataCreator dataCreator;
+		MyResourceDataCreator dataCreator;
+
+		sealed class MyResourceDataCreator : ResourceDataCreator {
+			public MyResourceDataCreator(ModuleDef module)
+				: base(module) {
+			}
+
+			protected override string GetAssemblyFullName(string simpleName) {
+				var asm = TheAssemblyResolver.Instance.Resolve(new AssemblyNameInfo(simpleName), Module);
+				return asm == null ? null : asm.FullName;
+			}
+		}
 
 		public ResourceConverter(ModuleDefMD module, ResourceInfo[] infos) {
 			this.module = module;
-			this.dataCreator = new ResourceDataCreator(module);
+			this.dataCreator = new MyResourceDataCreator(module);
 			this.infos = infos;
 		}
 
@@ -65,7 +78,7 @@ namespace de4dot.code.deobfuscators.CodeVeil {
 				break;
 
 			case 4:		// char[]
-				resourceData = dataCreator.Create(reader.ReadChars(info.length));
+				resourceData = new CharArrayResourceData(dataCreator.CreateUserResourceType(CharArrayResourceData.ReflectionTypeName), reader.ReadChars(info.length));
 				break;
 
 			case 5:		// sbyte
@@ -125,11 +138,11 @@ namespace de4dot.code.deobfuscators.CodeVeil {
 				break;
 
 			case 19:	// Icon
-				resourceData = dataCreator.CreateIcon(reader.ReadBytes(info.length));
+				resourceData = new IconResourceData(dataCreator.CreateUserResourceType(IconResourceData.ReflectionTypeName), reader.ReadBytes(info.length));
 				break;
 
 			case 20:	// Image
-				resourceData = dataCreator.CreateImage(reader.ReadBytes(info.length));
+				resourceData = new ImageResourceData(dataCreator.CreateUserResourceType(ImageResourceData.ReflectionTypeName), reader.ReadBytes(info.length));
 				break;
 
 			case 31:	// binary
@@ -145,6 +158,60 @@ namespace de4dot.code.deobfuscators.CodeVeil {
 				Name = info.name,
 				ResourceData = resourceData,
 			};
+		}
+	}
+
+	class CharArrayResourceData : UserResourceData {
+		public static readonly string ReflectionTypeName = "System.Char[],mscorlib";
+		char[] data;
+
+		public CharArrayResourceData(UserResourceType type, char[] data)
+			: base(type) {
+			this.data = data;
+		}
+
+		public override void WriteData(BinaryWriter writer, IFormatter formatter) {
+			formatter.Serialize(writer.BaseStream, data);
+		}
+
+		public override string ToString() {
+			return string.Format("char[]: Length: {0}", data.Length);
+		}
+	}
+
+	class IconResourceData : UserResourceData {
+		public static readonly string ReflectionTypeName = "System.Drawing.Icon,System.Drawing";
+		Icon icon;
+
+		public IconResourceData(UserResourceType type, byte[] data)
+			: base(type) {
+			icon = new Icon(new MemoryStream(data));
+		}
+
+		public override void WriteData(BinaryWriter writer, IFormatter formatter) {
+			formatter.Serialize(writer.BaseStream, icon);
+		}
+
+		public override string ToString() {
+			return string.Format("Icon: {0}", icon);
+		}
+	}
+
+	class ImageResourceData : UserResourceData {
+		public static readonly string ReflectionTypeName = "System.Drawing.Bitmap,System.Drawing";
+		Bitmap bitmap;
+
+		public ImageResourceData(UserResourceType type, byte[] data)
+			: base(type) {
+			bitmap = new Bitmap(Image.FromStream(new MemoryStream(data)));
+		}
+
+		public override void WriteData(BinaryWriter writer, IFormatter formatter) {
+			formatter.Serialize(writer.BaseStream, bitmap);
+		}
+
+		public override string ToString() {
+			return "Bitmap";
 		}
 	}
 }
