--- de4dot.code/resources/ResourceElement.cs
@@ -1,29 +1,0 @@
-﻿/*
-    Copyright (C) 2011-2014 de4dot@gmail.com
-
-    This file is part of de4dot.
-
-    de4dot is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    de4dot is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with de4dot.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-namespace de4dot.code.resources {
-	public class ResourceElement {
-		public string Name { get; set; }
-		public IResourceData ResourceData { get; set; }
-
-		public override string ToString() {
-			return string.Format("N: {0}, V: {1}", Name, ResourceData);
-		}
-	}
-}
