--- de4dot.code/resources/ResourceTypeCode.cs
@@ -1,43 +1,0 @@
-﻿/*
-    Copyright (C) 2011-2014 de4dot@gmail.com
-
-    This file is part of de4dot.
-
-    de4dot is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    de4dot is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with de4dot.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-namespace de4dot.code.resources {
-	public enum ResourceTypeCode {
-		Null,
-		String,
-		Boolean,
-		Char,
-		Byte,
-		SByte,
-		Int16,
-		UInt16,
-		Int32,
-		UInt32,
-		Int64,
-		UInt64,
-		Single,
-		Double,
-		Decimal,
-		DateTime,
-		TimeSpan,
-		ByteArray = 0x20,
-		Stream = 0x21,
-		UserTypes = 0x40,
-	}
-}
